This commit introduces the initial implementation of Services and Repositories for the Orders, Pizzas, and Ingredients entities.
